+++ digestbot/core/common/Enums.py
@@ -1,13 +1,5 @@
 from enum import Enum, unique
 
-# TODO: fix it
-# TODO: maybe https://stackoverflow.com/questions/37183612/how-to-define-a-mapping-of-enum-members-in-the-enum-type?
-SORTING_TYPE_MAPPER = {
-    "replies": "reply_count",
-    "length": "thread_length",
-    "reactions": "reactions_rate",
-}
-
 
 @unique
 class SortingType(Enum):
