+++ digestbot/core/ui_processor/common.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class UserRequest:
+    text: str  # user text
+    user: str  # author's username
+    channel: str  # ID of channel from Slack
+    ts: str  # timestamp of the message
+    is_im: bool  # is it private message to bot (or in public channel)
